requirements.txt
Django==2.2.6
djangorestframework==3.10.3
freeze==1.0.10
Pillow==6.2.1
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
